Refactor server actions validation

+++ src/actions/actions.ts
@@ -1,17 +1,23 @@
 'use server';
 
 import prisma from '@/lib/db';
-import { PetEssentials } from '@/lib/types';
 import { sleep } from '@/lib/utils';
-import { Pet } from '@prisma/client';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 import { revalidatePath } from 'next/cache';
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
+  const validatedPet = petFormSchema.safeParse(pet); // Validate the pet data against the schema
+
+  if (!validatedPet.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -22,13 +28,21 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], pet: PetEssentials) {
+export async function editPet(petId: unknown, pet: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(pet);
+
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -38,11 +52,17 @@ export async function editPet(petId: Pet['id'], pet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: 'Invalid pet ID',
+    };
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
